requirements.txt
pytest>=6.2.5
requests>=2.28.1
allure-pytest>=2.13.2
Faker>=19.13.0